[Fix] Frame number to 16bit
PutVideoFrameHeader, CODEC_TAG_FRAME_NUMBER need 16bit value
Overflowing frame_number doesn't seem to be a problem, used for debug purposes

+++ Codec/encoder.h
@@ -327,7 +327,7 @@ typedef struct encoder {	// Encoder state (derived from codec)
 	int chromaFullRes;		// True if the encoder was initialized with full resolution chroma
 
 	// Number of the most recent frame processed by the encoder (first frame is number one)
-	uint32_t frame_number;
+	uint16_t frame_number;
 
 #if _THREADED_ENCODER
